+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## mu-5.0.2-03
+
+Release date: xx/09/2025
+
+* Certification allocation conflicts are now handled gracefully.
+
 ## mu-5.0.2-02
 
 Release date: 31/08/2025